Rename example packages
